chore: refactor helpers
