+++ src/routes/dashboard/+page.server.ts
@@ -4,16 +4,10 @@ import { superValidate, setError, type SuperValidated, type Infer } from 'svelte
 import { type FormSchema, formSchema } from '../schema';
 import { type EditFormSchema, editFormSchema } from './schema';
 import { zod } from 'sveltekit-superforms/adapters';
-import {
-	createLink,
-	deleteLink,
-	getLinks,
-	getRedirectsOfLinks,
-	type LinkCreate,
-	type LinkRead
-} from '$lib/api/links';
+import { client } from '$lib/client';
 import crypto from 'crypto';
-import { type BannerRead, createBanner } from '$lib/api/banners';
+import { createLink } from '$lib/helpers/createLink';
+import { fetchRedirects } from '$lib/helpers/fetchRedirects';
 export interface LinkData extends SuperValidated<Infer<FormSchema>> {
 	url: string | null;
 }
@@ -22,11 +16,14 @@ export const load: ServerLoad = async (event) => {
 	const token = await event.cookies.get('access_token');
 	if (!token) redirect(302, '/login');
 
-	const links = await getLinks(token);
-	if (links === 401) {
-		redirect(302, '/login');
-	}
-	if (links === 404) {
+	const links = await client.links
+		.listLinks({ headers: { authorization: `Bearer ${token}` } })
+		.then((res) => {
+			if (res.status === 200) return res.body;
+			if (res.status === 401) redirect(302, '/login');
+			return null;
+		});
+	if (!links) {
 		return {
 			links: [],
 			urlForm: await superValidate(zod(formSchema)),
@@ -34,9 +31,9 @@ export const load: ServerLoad = async (event) => {
 			url: null
 		};
 	}
-	let linksWithVisits = await getRedirectsOfLinks(links, token);
+	const linksWithRedirects = await fetchRedirects(links, token);
 	return {
-		links: linksWithVisits,
+		links: linksWithRedirects,
 		urlForm: await superValidate(zod(formSchema)),
 		editForm: await superValidate(zod(editFormSchema)),
 		url: null,
@@ -45,94 +42,7 @@ export const load: ServerLoad = async (event) => {
 };
 
 export const actions: Actions = {
-	createLink: async (event) => {
-		const form = await superValidate(event, zod(formSchema));
-		if (!form.valid) {
-			return fail(400, {
-				form,
-				url: null
-			});
-		}
-
-		const token = event.cookies.get('access_token');
-		if (!token) redirect(302, '/login');
-		const banner = event.cookies.get('banner_id');
-		if (!banner) {
-			const bannerCreate: BannerRead = await createBanner(
-				{
-					title: `${crypto.randomUUID().toString()}`,
-					description: `${crypto.randomUUID().toString()}`
-				},
-				token
-			);
-			event.cookies.set('banner_id', bannerCreate.id, {
-				path: '/',
-				sameSite: 'strict',
-				httpOnly: false
-			});
-		}
-		const linkRequest: LinkCreate = {
-			short_id: form.data.short_id,
-			redirect_url: form.data.redirect_url,
-			banner_id: event.cookies.get('banner_id')!
-		};
-
-		if (form.data.expiration_date_enabled) {
-			if (!form.data.expiration_date) {
-				setError(form, 'expiration_date', 'Please specify the expiration date');
-				return fail(400, {
-					form,
-					url: null
-				});
-			}
-			linkRequest.expiration_date = form.data.expiration_date;
-		}
-		if (form.data.redirects_limit_enabled) {
-			if (!form.data.redirects_limit) {
-				setError(form, 'redirects_limit', 'Please specify the redirects limit');
-				return fail(400, {
-					form,
-					url: null
-				});
-			}
-			linkRequest.redirects_limit = parseInt(form.data.redirects_limit);
-		}
-
-		const linkResponse: LinkRead | null = await createLink(linkRequest, token);
-
-		if (!linkResponse) {
-			setError(form, 'short_id', 'This link already in use. Please specify other custom link');
-			return fail(400, {
-				form,
-				url: null
-			});
-		}
-		if (!linkResponse.owner_id) {
-			setError(form, 'redirect_url', '');
-			return fail(400, {
-				form,
-				url: null
-			});
-		}
-
-		const links = await getLinks(token);
-		if (links === 401) {
-			redirect(302, '/login');
-		}
-		if (links === 404) {
-			return {
-				links: [],
-				form: await superValidate(zod(formSchema)),
-				url: null
-			};
-		}
-		let linksWithVisits = await getRedirectsOfLinks(links, token);
-		return {
-			links: linksWithVisits,
-			form,
-			url: linkResponse.short_id
-		};
-	},
+	createLink: async (event) => await createLink(event, zod),
 	deleteLink: async (event) => {
 		const token = event.cookies.get('access_token');
 		if (!token) redirect(302, '/login');
